+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Extend compatibility down to VS Code 1.99.0.
 - Rename `erlab.manager` to `erlab.itool`.
 - Add `erlab.itool.useManager` setting to control ImageTool manager usage.
+- Show a notebook cell status bar button to open DataArray variables in ImageTool.
 
 ## [v0.0.2] - 2026-01-10